requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
colorama==0.4.6
coverage==7.8.0
Django==4.2.9
django-debug-toolbar==5.2.0
django-redis==5.4.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
Faker==37.3.0
inflection==0.5.1
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
pytest==8.3.5
pytest-cov==6.1.1
pytest-django==4.11.1
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
redis==6.1.0
sqlparse==0.5.3
tqdm==4.67.1
tzdata==2025.2
uritemplate==4.1.1
ruff==0.11.10
